Removed sci_base.js and a couple small tweaks
Removed sci_base.js as that is now called sci_general.js.  Made a couple of small tweaks to the sci_canvas_controls.js atom.  One had to do with 'mouseout' not clearing the button hover color on sciCanvasButton.  The other had to do with a bug in the horiz text centering for sciCanvasButton.