Update time_select_calendar.js
Fixed bugs in the capability to drop some of the time fields (minutes, hours or days) based on a seconds resolution value.  I was letting the 'this.num_fields' variable do all the work, but I needed to add some if/them logic to handle the fields that don't get used.